move side functions to helpers.go
